gr-rassadnikov: [refactor] Rename app
