+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="ru.ok.itmo.example">
+    package="ru.ok.itmo.tamtam">
 
     <uses-feature
         android:name="android.hardware.camera"
@@ -22,7 +22,7 @@
         android:theme="@style/Theme.Itmoexample"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="ru.ok.itmo.tamtam.MainActivity"
             android:exported="true"
             android:windowSoftInputMode="adjustResize">
             <intent-filter>
