+++ app/src/main/java/ru/ok/itmo/tamtam/client/LoginApi.kt
@@ -1,10 +1,10 @@
-package ru.ok.itmo.example.client
+package ru.ok.itmo.tamtam.client
 
 import retrofit2.Response
 import retrofit2.Retrofit
 import retrofit2.http.POST
 import retrofit2.http.Body
-import ru.ok.itmo.example.dto.UserAuthorization
+import ru.ok.itmo.tamtam.dto.UserAuthorization
 
 interface LoginApi {
     @POST("login")
