+++ app/src/main/java/ru/ok/itmo/tamtam/domain/ErrorType.kt
@@ -1,4 +1,4 @@
-package ru.ok.itmo.example.domain
+package ru.ok.itmo.tamtam.domain
 
 sealed class ErrorType : Exception() {
     class Unknown : ErrorType()
