+++ app/src/main/java/ru/ok/itmo/tamtam/ui/LoginFragment.kt
@@ -1,4 +1,4 @@
-package ru.ok.itmo.example.ui
+package ru.ok.itmo.tamtam.ui
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -8,12 +8,12 @@ import android.view.ViewGroup
 import androidx.core.widget.addTextChangedListener
 import androidx.fragment.app.viewModels
 import androidx.navigation.Navigation
-import ru.ok.itmo.example.domain.ErrorType
-import ru.ok.itmo.example.R
-import ru.ok.itmo.example.databinding.FragmentLoginBinding
-import ru.ok.itmo.example.domain.LoginState
-import ru.ok.itmo.example.domain.LoginViewModel
-import ru.ok.itmo.example.util.TextPresentObjects
+import ru.ok.itmo.tamtam.domain.ErrorType
+import ru.ok.itmo.tamtam.R
+import ru.ok.itmo.tamtam.databinding.FragmentLoginBinding
+import ru.ok.itmo.tamtam.domain.LoginState
+import ru.ok.itmo.tamtam.domain.LoginViewModel
+import ru.ok.itmo.tamtam.util.TextPresentObjects
 
 class LoginFragment : Fragment() {
     private val viewModel by viewModels<LoginViewModel>()
